useModalTrigger to ts
